// src/app/charts/chart-types.enum.ts
/**
 * Разработанные функции
 * @readonly
 */
export enum ChartTypes {
  /** Линейная функция `y = kx + b` */
  linear = 'linear',
  /** Парабола `y = x^2 + kx + b` */
  parabole = 'parabole',
  /** Гипербола `y = k/x` */
  hyperbole = 'hyperbole',
}